데이터매니저 구현중

+++ TextJRPG/Player.cpp
@@ -34,15 +34,15 @@ void Player::SetMonster(int _level)
 
 	if (level < 5)
 	{
-		mItem = Item("hp포션(소)", 10, 100);
+		mItem = Item("hp포션(소)", Potion, 10, 100);
 	}
 	else if (level >=5 && level < 10)
 	{
-		mItem = Item("hp포션(중)", 30, 200);
+		mItem = Item("hp포션(중)", Potion, 30, 200);
 	}
 	else if(level >=10)
 	{
-		mItem = Item("hp포션(대)", 50, 300);
+		mItem = Item("hp포션(대)", Potion, 50, 300);
 	}
 }
 
@@ -141,8 +141,9 @@ int Player::GetCurMp()
 	return curMp;
 }
 
-void Player::SetAtk(int)
+void Player::SetAtk(int _atk)
 {
+	atk += _atk;
 }
 
 int Player::GetAtk()
@@ -150,14 +151,24 @@ int Player::GetAtk()
 	return atk;
 }
 
-void Player::SetDef(int)
+void Player::SetDef(int _def)
 {
-
+	def += _def;
 }
 
 int Player::GetDef()
 {
 	return def;
+}
+
+void Player::SetGold(int _gold)
+{
+	gold += _gold;
+}
+
+int Player::GetGold()
+{
+	return gold;
 }
 
 bool Player::Death()
@@ -186,14 +197,29 @@ Item Player::GetItem()
 	return mItem;
 }
 
-Inventory Player::GetInventory()
+bool Player::BuyItem(Item _item)
 {
-	return inventory;
+	cout << "현재 소지금 : " << gold <<endl
+		<< "아이템 가격 : " << _item.GetPrice() << endl;
+	if (gold >= _item.GetPrice())
+	{	
+		gold -= _item.GetPrice();
+		cout << "구매를 완료 했습니다. " << endl
+			<< "현재 소지금 : " << gold << endl;
+		return true;
+	}
+	else
+	{
+		cout << "소지금이 부족해서 구매를 할 수 없습니다." << endl;
+		return false;
+	}
+
 }
 
 
-void Player::ReleaseItem()
+Inventory Player::GetInventory()
 {
+	return inventory;
 }
 
 
@@ -207,6 +233,7 @@ void Player::UseItem()
 	mItem = inventory.UseItem(stoi(num));
 	inventory.DeleteItem(stoi(num));
 
+	// =====아이템 능력치 구현=====
 	curHp += mItem.GetValue();
 	if (curHp >= maxHp)
 	{
@@ -224,12 +251,16 @@ void Player::SellItem()
 	}
 	else
 	{
-		cout << "판매할 아이템의 번호를 입력하세요." << endl;
+		cout << "아이템의 번호를 입력하세요." << endl;
 		string num;
 		cin >> num;
 
-		mItem = inventory.SellItem(stoi(num));
+		mItem = inventory.UseItem(stoi(num));
 		inventory.DeleteItem(stoi(num));
+
+		gold += mItem.GetPrice();
+		cout << "현재 소지금 : " << GetGold() << endl;
+		
 	}
 }
 
@@ -237,10 +268,10 @@ void Player::SellItem()
 
 void Player::LoadInventory()
 {
-	cout << "===== 장비 =====" << endl;
+	cout << "********** 장비 *********" ;
 	equipment.LoadInventory();
 
-	cout << "===== 인벤토리 =====" << endl;
+	cout << "******* 인벤토리 *******";
 	inventory.LoadInventory();
 	
 }
@@ -259,8 +290,62 @@ void Player::SetEquipment()
 		cin >> num;
 
 		mItem = inventory.UseItem(stoi(num));
-		inventory.DeleteItem(stoi(num));
-		equipment.SetEquipment(mItem);
+
+		if (equipment.SetEquipment(mItem))
+		{
+			// 장착에 성공하면 코드 실행
+			inventory.DeleteItem(stoi(num));
+
+			// ==== 장비 능력치 구현 ====
+
+			if (mItem.GetType() == 1)
+			{
+				atk += mItem.GetValue();
+			}
+			else if (mItem.GetType() == 2)
+			{
+				def += mItem.GetValue();
+			}
+			else if (mItem.GetType() == 3)
+			{
+				maxHp += mItem.GetValue();
+				curHp = maxHp;
+			}
+		}
+	}
+}
+
+void Player::ReleaseEquipment()
+{
+
+	if (equipment.EmptyInventory())
+	{
+		cout << "장착한 아이템이 없어 탈착할 수 없습니다." << endl;
+	}
+	else
+	{
+		cout << "탈착할 아이템의 번호를 입력하세요." << endl;
+		string num;
+		cin >> num;
+
+		mItem = equipment.UseItem(stoi(num));
+		equipment.DeleteItem(stoi(num));
+		inventory.ReleasEquipment(mItem);
+
+		// ==== 장비 능력치 구현 ====
+		if (mItem.GetType() == 1)
+		{
+			atk -= mItem.GetValue();
+		}
+		else if (mItem.GetType() == 2)
+		{
+			def -= mItem.GetValue();
+		}
+		else if (mItem.GetType() == 3)
+		{
+			maxHp -= mItem.GetValue();
+			curHp = maxHp;
+		}
 	}
 }
